+++ pgqueuer/qb.py
@@ -1,12 +1,17 @@
-"""Backward-compatibility shim. Canonical: pgqueuer.adapters.persistence.qb"""
+"""Backward-compatibility shim. Canonical locations:
+- Settings types: pgqueuer.domain.settings
+- SQL builders: pgqueuer.adapters.persistence.qb
+"""
 
 from pgqueuer.adapters.persistence.qb import (
-    DBSettings,
-    Durability,
-    DurabilityPolicy,
     QueryBuilderEnvironment,
     QueryQueueBuilder,
     QuerySchedulerBuilder,
+)
+from pgqueuer.domain.settings import (
+    DBSettings,
+    Durability,
+    DurabilityPolicy,
     add_prefix,
 )
